requirements.txt
tensorflow
numpy
Pillow
joblib
tkinter